requirements.txt
# Nongor Bot V3 - Premium Dependencies
# Python 3.8+ required

# Telegram Bot Framework
python-telegram-bot==21.0

# Google Gemini AI
google-generativeai==0.8.6

# PostgreSQL Driver
pg8000==1.30.3

# Web Scraping & HTTP
requests==2.31.0
beautifulsoup4==4.12.2

# Environment Configuration
python-dotenv==1.0.0

# Protocol Buffers (Gemini compatibility)
protobuf>=5.29.0

# ---- PREMIUM INTEGRATIONS ----

# Email (SendGrid) - Optional
sendgrid>=6.11.0

# Google Sheets Export - Optional
google-auth>=2.0.0
google-api-python-client>=2.0.0

# Charts & Reports - Optional
matplotlib>=3.7.0
